requirements.txt
# AnimeCut Serverless - Requirements Essenciais
# Versão funcional sem conflitos

# ==================== RUNPOD ====================
runpod>=1.5.0

# ==================== VIDEO PROCESSING ====================
moviepy==1.0.3
opencv-python-headless==4.10.0.84
imageio==2.34.0
imageio-ffmpeg==0.5.1
proglog==0.1.10

# ==================== IMAGE PROCESSING ====================
Pillow>=10.1.0
numpy==1.26.4

# ==================== AUDIO PROCESSING ====================
pydub>=0.25.1

# ==================== BACKBLAZE B2 ====================
boto3>=1.29.0

# ==================== UTILITIES ====================
requests>=2.31.0
tqdm>=4.66.0
